add error logging to student routes
